pyproject.toml
# Prompt Manager CLI - Project Configuration
#
# This pyproject.toml configures:
# - [tool.poetry]: Package metadata and dependencies (Poetry)
# - [tool.ruff]: Linting and formatting rules (Ruff)
# - [tool.mypy]: Static type checking configuration (mypy)
# - [tool.pytest.ini_options]: Test configuration (pytest)
#
# Note: Python >=3.12 is recommended, but >=3.11 is supported for compatibility

[tool.poetry]
name = "prompt-unifier"
version = "0.3.0"
description = "CLI tool for managing AI prompt templates with YAML frontmatter"
authors = ["Waewoo <waewoo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "prompt_unifier", from = "src"}]

[tool.poetry.scripts]
prompt-unifier = "prompt_unifier.cli.main:main"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
typer = "^0.20.0"
rich = "^14.0"
pydantic = "^2.5.0"
pyyaml = "^6.0.1"
gitpython = "^3.1.40"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
ruff = "^0.3.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
types-pyyaml = "^6.0.12"
detect-secrets = "^1.5.0"
bandit = "^1.8.6"
safety = "^3.7.0"
pip-audit = "^2.9.0"
commitizen = "^4.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Ruff configuration - Linting and formatting
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["prompt_unifier"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["SIM117"]  # Allow nested with statements in tests for clarity

# mypy configuration - Static type checking
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# pytest configuration - Testing
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = [
    "ignore:.*BaseCommand.*:DeprecationWarning:typer.*",
    "ignore:.*__version__.*:DeprecationWarning:typer.*",
]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
fail_under = 95
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",  # Exclude protocol ellipsis markers
]

# Bandit configuration - Security linting
[tool.bandit]
targets = ["src"]
exclude_dirs = ["/tests", "/agent-os"]
skips = []
severity = "medium"
confidence = "medium"

[tool.bandit.assert_used]
skips = ["*/test_*.py", "*/conftest.py"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.3.0"
version_files = [
    "pyproject.toml:version",
    "src/prompt_unifier/__init__.py:__version__"
]
tag_format = "v$version"
changelog_file = "CHANGELOG.md"
update_changelog_on_bump = true
bump_message = "bump: version $current_version → $new_version"